eLearning/requirements.txt
amqp==5.2.0
api-client==1.3.1
apiclient==1.0.4
asgiref==3.8.1
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-bootstrap5==24.2
django-crispy-forms==2.3
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==27.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
kombu==5.4.0
msgpack==1.0.8
passlib==1.7.4
pillow==10.4.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0
